ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic>=2.6.0
requests==2.31.0
aiofiles==23.2.1
httpx>=0.25.0 